refactor: 영화 상세정보 모달 모듈 분리
